Im on oblivion from the album Visions also lastfm in separate files and fixes and fixes and string fixes
