+++ src/locales/en-US/commands.json
@@ -44,9 +44,6 @@
   "github": {
     "commandDescription": "Searches for a user, repository or organization on Github",
     "commandUsage": "<type> <query>",
-    "noType": "Please insert a valid type",
-    "types": "Types",
-    "selectResult": "Please send the number of the result (10 seconds)",
     "subcommands": {
       "user": {
         "commandDescription": "Searches for an user on GitHub",
@@ -85,9 +82,6 @@
   "spotify": {
     "commandDescription": "Searches for a song, album, artist or a playlist on Spotify, or displays a user info.",
     "commandUsage": "<type> <query>",
-    "noTracksFound": "I couldn't find any tracks with that name.",
-    "noType": "Please insert a valid type",
-    "types": "Types",
     "album": "Album",
     "genres": "Genres",
     "artist": "Artist",
@@ -99,22 +93,15 @@
     "moreAlbums": "...and {{albums}} more albums",
     "followers": "Followers",
     "followersCount": "{{followers}} followers",
-    "selectResult": "Please send the number of the result (10 seconds)",
     "subcommands": {
       "track": {
         "commandDescription": "Searches for a song on Spotify",
         "commandUsage": "<song>",
-        "noTrack": "You need to give me a track name.",
-        "notFound": "I couldn't find any track with that name.",
-        "results": "Track results for {{query}}",
         "trackInfo": "Track information"
       },
       "album": {
         "commandDescription": "Searches for an album on Spotify",
         "commandUsage": "<album>",
-        "noAlbum": "You need to give me an album name.",
-        "notFound": "I couldn't find any album with that name.",
-        "results": "Album results for {{query}}",
         "albumInfo": "Album information",
         "albumType": "Album type",
         "types": {
@@ -125,25 +112,17 @@
       "artist": {
         "commandDescription": "Searches for an artist on Spotify",
         "commandUsage": "<artist>",
-        "noArtist": "You need to give me an artist name.",
-        "notFound": "I couldn't find any artist with that name.",
-        "results": "Artist results for {{query}}",
         "artistInfo": "Artist information"
       },
       "playlist": {
         "commandDescription": "Searches for a playlist on Spotify",
         "commandUsage": "<playlist>",
-        "noPlaylist": "You need to give me a playlist name.",
-        "notFound": "I couldn't find any playlist with that name.",
-        "results": "Playlist results for {{query}}",
         "playlistInfo": "Playlist information",
         "createdBy": "Created by"
       },
       "user": {
         "commandDescription": "Get information about a user on Spotify",
         "commandUsage": "<user>",
-        "noUser": "You need to give me a valid user.",
-        "notFound": "I couldn't find any users.",
         "userInfo": "User information"
       }
     }
@@ -914,37 +893,26 @@
   "lastfm": {
     "commandDescription": "Get an user information or search for a track, album or artist on Last.fm",
     "commandUsage": "<type> <query>",
-    "notFound": "The results for {{query}} returned 0 matches.",
     "topArtists": "Top artists (1 month)",
     "playcount": "Playcount",
     "playcountCount": "Played {{times}} times",
     "listeners": "Listeners",
     "album": "Album",
     "artist": "Artist",
-    "selectQuery": "Please send the number of the result (10 seconds)",
     "readMore": "Read more on Last.fm",
     "tags": "Tags",
     "tracks": "Tracks",
     "registered": "Registered in",
     "subcommands": {
       "track": {
-        "missing": "Please insert a track to search.",
-        "results": "Track results for {{query}}",
-        "resultsCount": "{{count}} tracks found",
         "commandDescription": "Search for a track on Last.fm",
         "commandUsage": "<track>"
       },
       "album": {
-        "missing": "Please insert an album to search.",
-        "results": "Album results for {{query}}",
-        "resultsCount": "{{count}} albums found",
         "commandDescription": "Search for an album on Last.fm",
         "commandUsage": "<album>"
       },
       "artist": {
-        "missing": "Please insert an artist to search",
-        "results": "Artist results for {{query}}",
-        "resultsCount": "{{count}} artists found",
         "commandDescription": "Search for an artist on Last.fm",
         "commandUsage": "<artist>"
       },
